backend/requirements.txt
Flask==2.3.3
flask-jwt-extended==4.6.0
SQLAlchemy==2.0.23
PyMySQL==1.1.0
python-dotenv==1.0.1
scikit-learn==1.3.2
nltk==3.8.1
transformers==4.37.2
sentence-transformers==2.3.1
twilio==8.10.0
apscheduler==3.10.4
Werkzeug==3.0.1

